backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.0

# Environment variables
python-decouple==3.8

# Geolocation
geopy==2.4.0
openrouteservice==2.3.3

# Image processing
Pillow==10.1.0

# Firebase notifications
fcm-django==2.0.0
firebase-admin==6.9.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==2.43.0

# Date/Time
python-dateutil==2.8.2

# API Documentation
drf-yasg==1.21.7
